working on Reschedule

+++ src/main/java/com/appbackend/example/AppBackend/models/RescheduleDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 
 @Data
 public class RescheduleDto {
-    private long disbursementId;
+    private int disbursementId;
     private LocalDate rescheduledDate;
